updated reaction component
